+++ packages/vite/src/node/plugins/importMetaGlob.ts
@@ -42,10 +42,16 @@ interface ParsedGeneralImportGlobOptions extends GeneralImportGlobOptions {
 }
 
 export function importGlobPlugin(config: ResolvedConfig): Plugin {
-  if (config.command === 'build' && config.nativePluginEnabledLevel >= 1) {
+  if (config.isBundled && config.nativePluginEnabledLevel >= 1) {
     return nativeImportGlobPlugin({
       root: config.root,
       restoreQueryExtension: config.experimental.importGlobRestoreExtension,
+      isV2:
+        config.nativePluginEnabledLevel >= 2
+          ? {
+              sourcemap: !!config.build.sourcemap,
+            }
+          : undefined,
     })
   }